Move NavigationInspector to new file
